1024 push study 핸들러 mariadb 적용 / guilder 핸들러 mriaddb 적용 (구현 완료 드디어)

+++ app-client/src/main/java/com/ogong/pms/handler/myStudy/guilder/WatingGuilderListHandler.java
@@ -2,6 +2,7 @@ package com.ogong.pms.handler.myStudy.guilder;
 
 import java.util.List;
 import com.ogong.pms.dao.StudyDao;
+import com.ogong.pms.domain.Guilder;
 import com.ogong.pms.domain.Member;
 import com.ogong.pms.domain.Study;
 import com.ogong.pms.handler.AuthPerMemberLoginHandler;
@@ -27,6 +28,17 @@ public class WatingGuilderListHandler implements Command {
     int inputNo = (int) request.getAttribute("inputNo");
 
     Study myStudy = studyDao.findByNo(inputNo);
+
+    // 해당 스터디에 구성원 목록 가져오기
+    List<Guilder> guilderList = studyDao.findByGuilderAll(myStudy.getStudyNo());
+    for (Guilder guilder : guilderList) {
+      if (guilder.getGuilderStatus() == 2) {
+        myStudy.getMembers().add(guilder.getMember());
+
+      } else if (guilder.getGuilderStatus() == 1) {
+        myStudy.getWatingMember().add(guilder.getMember());
+      }
+    }
 
     if (!myStudy.getWatingMember().isEmpty()) {
       for (Member m : myStudy.getWatingMember()) {
@@ -103,6 +115,13 @@ public class WatingGuilderListHandler implements Command {
           }
 
           if (watingMember.getPerNickname().equals(input)) {
+
+            String answer = Prompt.inputString(" 정말 거절하시겠습니까?(네 / 아니오) ");
+            if (!answer.equals("네")) {
+              System.out.println(" >> 거절을 취소하였습니다.");
+              return;
+            }
+
             studyDao.deleteGuilder(myStudy.getStudyNo(), watingMember.getPerNo());
             System.out.printf(" >> '%s님'의 구성원 신청이 거절되었습니다.\n", watingMember.getPerNickname());
             break;
